Support getting more than 50 YouTube entries
YouTube API at the moment only allows `max-results` of maximum 50
but there are cases we want to get more than that.

The end-user can specify `max-results` to the number of entries
wanted in the `params`, and this will parse into multiple requests
using the `start-index` param.

Added `async` as dependency so that all the requests can be done
in parallel.

Bump version up.