Add support for API v3
The YouTube Data API (v2) has been officially deprecated as of March 4, 2014. It will be removed one year after the announcement or April 20, 2015.

This commit adds a new method to work with API v3. It requires a `playlistId` in `params`, and can optionally take a `part` and `maxResults`. It recursively requests all (if `maxResults` is not set) or up to `maxResults` uploaded videos by authenticated user. The `playlistId` is the ID of the `uploads` playlist.