Fixed a bug of `params` being overriden
Previously, the `params.maxResults` was updated in every recursion to get all the videos. At the end, it had the last number of videos in the `params`. Due to Javascript's object is not referenced by value, this `maxResults` was kept in the API data.

Once the bundle was expired, SPAS went and made request again, but then `maxResults` was not the same as specified in bundle definition any more, thus providing wrong number of results.

This fix clones the `params` object in every recursion to keep the original value.